add fused_rope to functional
