+++ apex/transformer/functional/__init__.py
@@ -1,5 +1,11 @@
+from apex.transformer.functional.fused_rope import (
+    fused_apply_rotary_pos_emb,
+    fused_apply_rotary_pos_emb_cached,
+)
 from apex.transformer.functional.fused_softmax import FusedScaleMaskSoftmax
 
 __all__ = [
     "FusedScaleMaskSoftmax",
+    "fused_apply_rotary_pos_emb",
+    "fused_apply_rotary_pos_emb_cached",
 ]
